[#187] modify: Mileage 타입 Integer -> Double

+++ rest/src/main/java/com/wherecar/rest/car/application/dto/CarRegisterRequest.java
@@ -17,7 +17,7 @@ public class CarRegisterRequest {
     private String make;
     private String model;
     private String year;
-    private Integer mileage;
+    private Double mileage;
     private OwnerType ownerType;
     private AcquisitionType acquisitionType;
     private Long companyId;
