+++ rest/src/main/java/com/wherecar/rest/car/application/dto/CarResponse.java
@@ -19,7 +19,7 @@ public class CarResponse {
     private String make;
     private String model;
     private String year;
-    private Integer mileage;
+    private Double mileage;
     private OwnerType ownerType;
     private AcquisitionType acquisitionType;
     private String companyName;
